Ixx inertia calculation with 'y' Not 'Y'
Well, it seems like this one calculates X-axis Inertia Tensor. But than, it should add all of the 'small Mass' * 'Distance'^2 . 
However, in this code, it uses Y.*Y, but Y is a matrix with size of ((N/2)x(N/2)), and it contains data of from each section [0,b/2-delta], how much OUTER(closer to wing tip) discrete points[delta/2~b/2-delta/2] are. And so it has that unique size. 

So I think, for calculating the correct Ixx, it should use 'y', since that is just a representation of how far each discrete points are from 'CENTER'. Which also equals to : Y(1). Current code outputs Way bigger Inertia than just 'y.*y*(small Mass)' . For ex) [0.2825 vs 0.047].

Is this intended for any purpose ? Cuz It really doesn't make physics-sense to me.