saving current 'rd' to 'rd_last'
'rd' is 'Weight Reduction Ratio'. It is used in line #872. I don't fully understand why in that case, 'weight' would be Fluctuating.
Anyway, the loop doesn't save 'rd_last' variable in any part. 
And it is initialized as 1e+60. Because of that, 'if' statement is Never True.
Therefore I added a line to save last 'rd' data into 'rd_last'.